move all parsing logic to parse.go for both azure operators
