+++ operator/builtin/input/azure/eventhub/event_hub.go
@@ -4,11 +4,9 @@ import (
 	"context"
 	"fmt"
 
-	azhub "github.com/Azure/azure-event-hubs-go/v3"
 	"github.com/observiq/stanza/operator"
 	"github.com/observiq/stanza/operator/builtin/input/azure"
 	"github.com/observiq/stanza/operator/helper"
-	"go.uber.org/zap"
 )
 
 const operatorName = "azure_event_hub_input"
@@ -109,24 +107,4 @@ func (e *EventHubInput) Start() error {
 // Stop will stop generating logs.
 func (e *EventHubInput) Stop() error {
 	return e.StopConsumers()
-}
-
-// handleEvent handles an event recieved by an Event Hub consumer.
-func (e *EventHubInput) handleEvent(ctx context.Context, event *azhub.Event) error {
-	e.WG.Add(1)
-	defer e.WG.Done()
-
-	entry, err := e.NewEntry(nil)
-	if err != nil {
-		e.Errorw("", zap.Error(err))
-		return err
-	}
-
-	if err := azure.ParseEvent(*event, entry); err != nil {
-		e.Errorw("", zap.Error(err))
-		return err
-	}
-
-	e.Write(ctx, entry)
-	return nil
 }
